requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
dlib==19.17.0
Flask==1.1.1
gcloud==0.17.0
googleapis-common-protos==1.6.0
httplib2==0.13.1
idna==2.8
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
jws==0.1.3
MarkupSafe==1.1.1
numpy==1.17.0
oauth2client==3.0.0
opencv-contrib-python==4.1.0.25
Pillow==6.1.0
pip==19.0.3
protobuf==3.9.1
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycryptodome==3.4.3
pyfacy==1.0.1
pyfacy-dlib-models==0.0.4
Pyrebase==3.0.27
python-jwt==2.0.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.0
scikit-learn==0.21.3
scipy==1.3.1
setuptools==40.8.0
six==1.12.0
sklearn==0.0
urllib3==1.25.3
Werkzeug==0.15.5
